+++ lib/claide/command.rb
@@ -256,6 +256,28 @@ module CLAide
       end
     end
 
+    # Adds a new option for the current command.
+    #
+    # This method can be used in conjunction with overriding `options`.
+    #
+    # @return [void]
+    #
+    # @example
+    #
+    #   option '--help', 'Print help banner '
+    #
+    def self.option(name, description)
+      mod = Module.new do
+        define_method(:options) do
+          [
+            [name, description],
+          ].concat(super())
+        end
+      end
+      extend(mod)
+    end
+    private_class_method :option
+
     # Handles root commands options if appropriate.
     #
     # @param  [ARGV] argv